refactor: move functions-only helpers from column_helpers to functions_helpers
_md5_sparklike, _re_split_sparklike, _now_batch, and _substring_sparklike
are only used by functions.py — move them to functions_helpers.py where
they belong per the project's helpers convention.
